Notifications (Additions and Changes)
Summary:
Changed obj_notification so that after a certain amount of time
(currently randomly chosen between 2 to 5 seconds), it will be removed
and a bullet will be spawned in its place. That bullet will only go up
at a given speed (currently 30).
FIxed obj_spawn to accomodate for the changes made on spr_notification.

Added:
-spr_bullet
-fnt_Notification_Nums
-obj_counter
-obj_countdown
-obj_notification_nums
-obj_bullet
Changed:
-spr_notification
-obj_notification
-Falling_Room